+++ ExamenLab5P2_MichaelMartinez/src/empleado.java
@@ -132,7 +132,7 @@ public class empleado extends usuarios  {
             id="0000-";
         }
         id+=nacimiento.getYear()+"-";
-        double ultimo=random.nextDouble(10000,40000);
+        int ultimo=random.nextInt(10000,40000);
         id+=ultimo;
         return id;
     }
